requirements.txt
# Core Django & REST Framework
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.7.0
django-ratelimit==4.1.0
django-allauth==65.12.0
asgiref==3.8.1
apscheduler==3.10.4

# Database
psycopg==3.2.6
psycopg2==2.9.10

# Email & Authentication
PyJWT==2.10.1
cryptography==46.0.2

# Environment & Configuration
python-dotenv==1.0.1
whitenoise==6.6.0

# HTTP & Requests
requests==2.31.0
twilio==9.3.4
urllib3==2.0.3
certifi==2023.11.17
charset-normalizer==3.1.0
idna==3.4

# Geocoding & Location Services
geopy==2.4.1
geographiclib==2.0

# File Processing & Export
openpyxl==3.1.2
XlsxWriter==3.1.2
python-pptx==0.6.21
Pillow==10.3.0
python-bidi==0.4.2

# Video & Audio Generation
gTTS==2.5.4
moviepy==2.2.1
imageio==2.37.2
imageio-ffmpeg==0.6.0
proglog==0.1.12

# Data Processing
pandas==2.2.0

# Progress & Logging
tqdm==4.67.1

# Testing
pytest==8.4.2
pytest-django==4.11.1



# Transitive dependencies (kept for compatibility)
click==8.1.7
colorama==0.4.6
attrs==23.2.0
six==1.16.0
typing_extensions==4.9.0
packaging==23.1
decorator==5.2.1
blinker==1.7.0
MarkupSafe==2.1.3
Jinja2==3.1.2
filelock==3.18.0
